Move reports into it's own directory
